Always use square grid cells
Determine grid cell dimensions by the minimum of the screen width /
number of columns and the screen height / number of rows
